test(e2e): extract waitForConfigPage helper for dc-config hydration wait
add-source.spec.ts's openConfigPage() inlined the same "wait for
dc-config to hydrate" boilerplate (waitForSelector + waitForFunction
checking textContent doesn't include 'Loading') already duplicated in
config.spec.ts, mirroring the waitForTaskList/waitForTaskDetail/
waitForRunDetail pattern already in helpers/webui.ts. Add
waitForConfigPage(page) there and use it from both specs so there's
one place to update; config.spec.ts's separate "server settings
section" assertion (a distinct content-length check) is left as-is
since it isn't the same boilerplate.
